+++ Sistema-Ofi-Me/src/main/java/com/sistemaOficinaMecanica/repository/ServicoRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.sistemaOficinaMecanica.model.Servico;
 
-public interface ServicoRepository  extends JpaRepository<Servico, Integer>{
+public interface ServicoRepository  extends JpaRepository<Servico, Long>{
